Use Value Object for Azure Application Identity
Use a Value Object for the Azure resources instead of using
a hash object.

This commit also restructures the `call` function of
`validate_application_identity` to better tell the story
and to align between this class and the corresponding one
of authn-k8s.

It also removes the redundant class ValidateAzureAnnotations
and performs the validations as part of the general class